// src/lib/types/extendedDossier.type.ts
export interface extendDemande {
    equipement: string;
    modele: string;
    marque: string;
    fabricant: string;
    type: string;
    description: string;
    quantiteEquipements: string;
    fiche_technique : File;
}

export interface extendDossier {
    libelle : string
    courrier : File
}